+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - **GitLab API** extract_upstream_url_information can now also handle KDE URL's
 - **Package Handler, GitLab API, Test** KDE packages now use the GitLab REST API instead of the old playwright implementation
 - **Package Handler** Don't throw an exit when the package was found in multiple repositories, return None instead
+- **GitLab API** Remove the extraction of the creation date of package tags since it is currently not relevant and not used
 
 ### Bug fixes